+++ app/javascript/vue/tasks/collection_objects/filter/components/filters/geographic.vue
@@ -37,7 +37,7 @@
     </div>
     <div v-else>
       <georeference-map
-        width="330px"
+        width="100%"
         height="300px"
         ref="leaflet"
         :geojson="geojson"
@@ -51,7 +51,6 @@
         :tiles-selection="false"
         :edit-mode="false"
         :zoom="1"
-        :tooltips="false"
         @geoJsonLayerCreated="addShape"
       />
     </div>