+++ app/javascript/vue/tasks/collection_objects/simple_new_specimen/store/actions/actionNames.js
@@ -1,0 +1,10 @@
+export default {
+  CreateCollectingEvent: 'createCollectingEvent',
+  CreateCollectionObject: 'createCollectionObject',
+  CreateIdentifier: 'createIdentifier',
+  CreateNewSpecimen: 'createNewSpecimen',
+  CreateTaxonDetermination: 'createTaxonDetermination',
+  GetIdentifiers: 'getIdentifiers',
+  GetRecent: 'getRecent',
+  ResetStore: 'resetStore'
+}