+++ app/javascript/vue/tasks/collection_objects/simple_new_specimen/store/actions/createNewSpecimen.js
@@ -1,0 +1,37 @@
+import ActionNames from './actionNames.js'
+
+export default async function () {
+  const promises = []
+
+  if (
+    !this.createdCE &&
+    (
+      this.collectingEvent.verbatim_label ||
+      this.collectingEvent.verbatim_locality ||
+      this.geographicArea
+    )
+  ) {
+    await this[ActionNames.CreateCollectingEvent]()
+  }
+
+  this.createdCO = (await this[ActionNames.CreateCollectionObject]()).body
+
+  if (
+    this.identifier &&
+    this.namespace &&
+    !this.createdIdentifiers.length
+  ) {
+    promises.push(this[ActionNames.CreateIdentifier]())
+  }
+
+  if (this.otu) {
+    promises.push(this[ActionNames.CreateTaxonDetermination]())
+  }
+
+  Promise.allSettled(promises).then(_ => {
+    this[ActionNames.GetRecent]()
+    this[ActionNames.ResetStore]()
+
+    TW.workbench.alert.create('New specimen was successfully created', 'notice')
+  })
+}