+++ app/javascript/vue/tasks/collection_objects/stepwise/determinations/app.vue
@@ -16,10 +16,12 @@
     <div class="horizontal-left-content align-start">
       <div class="full_width">
         <LabelList class="margin-medium-bottom" />
-        <CollectionObjectList />
+        <CollectionObjectList v-if="selectedLabel" />
       </div>
 
-      <div class="margin-medium-left">
+      <div
+        id="right-column"
+        class="margin-medium-left">
         <CuttoffInput class="margin-medium-bottom" />
         <TaxonDetermination />
       </div>
@@ -38,7 +40,7 @@
 </template>
 
 <script setup>
-import { ref } from 'vue'
+import { ref, watch } from 'vue'
 import TaxonDetermination from './components/TaxonDetermination.vue'
 import CollectionObjectList from './components/CollectionObject/CollectionObjectList.vue'
 import VBtn from 'components/ui/VBtn/index.vue'
@@ -56,6 +58,8 @@ const {
   createDeterminations,
   collectionObjects,
   taxonDetermination,
+  loadCollectionObjects,
+  ghostCount,
   selectedCOIds
 } = useStore()
 
@@ -78,6 +82,19 @@ const handleClick = async () => {
   }
 }
 
+watch(
+  selectedLabel,
+  label => {
+    if (label) {
+      loadCollectionObjects(1).then(_ => {
+        if (ghostCount.value) {
+          TW.workbench.alert.create(`Warning, ${ghostCount.value} additional specimens identical except for whitepace are included`, 'notice')
+        }
+      })
+    }
+  }
+)
+
 </script>
 
 <script>
@@ -85,3 +102,9 @@ export default {
   name: 'StepwiseDeterminations'
 }
 </script>
+
+<style scoped>
+#right-column {
+  width: 600px;
+}
+</style>