+++ app/models/topic.rb
@@ -1,4 +1,4 @@
-# A Topic is a user defined subject.  It is used in conjuction with a citation on an OTU.  
+# A Topic is a user defined subject.  It is used in conjuction with a citation on an OTU.
 # Topics assert that "this source says something about this taxon on this topic."
 #
 class Topic < ControlledVocabularyTerm
@@ -43,11 +43,11 @@ class Topic < ControlledVocabularyTerm
     p = Topic.arel_table
 
     # i is a select manager
-    i = t.project(t['topic_id'], t['created_at']).from(t)
-      .where(t['created_at'].gt(10.weeks.ago))
-      .where(t['created_by_id'].eq(user_id))
+    i = t.project(t['topic_id'], t['updated_at']).from(t)
+      .where(t['updated_at'].gt(10.weeks.ago))
+      .where(t['updated_by_id'].eq(user_id))
       .where(t['project_id'].eq(project_id))
-      .order(t['created_at'].desc)
+      .order(t['updated_at'].desc)
 
     # z is a table alias
     z = i.as('recent_t')