+++ lib/export/coldp/files/taxon.rb
@@ -11,10 +11,10 @@
 module Export::Coldp::Files::Taxon
 
   IRI_MAP = {
-    extinct: 'https://api.catalogue.life/datapackage#Taxon.extinct',                         # 1,0
-    temporal_range_end: 'https://api.catalogue.life/datapackage#Taxon.temporal_range_end',   # from https://api.catalogue.life/vocab/geotime
-    temporal_range_start: 'https://api.catalogue.life/datapackage#Taxon.temporal_range_end', # from https://api.catalogue.life/vocab/geotime
-    lifezone: 'https://api.catalogue.life/datapackage#Taxon.lifezone',                       # from https://api.catalogue.life/vocab/lifezone
+    extinct: 'https://api.checklistbank.org/datapackage#Taxon.extinct',                         # 1,0
+    temporal_range_end: 'https://api.checklistbank.org/datapackage#Taxon.temporal_range_end',   # from https://api.checklistbank.org/vocab/geotime
+    temporal_range_start: 'https://api.checklistbank.org/datapackage#Taxon.temporal_range_end', # from https://api.checklistbank.org/vocab/geotime
+    lifezone: 'https://api.checklistbank.org/datapackage#Taxon.lifezone',                       # from https://api.checklistbank.org/vocab/lifezone
   }
 
   SKIPPED_RANKS = %w{
@@ -41,6 +41,15 @@ module Export::Coldp::Files::Taxon
     # * if two OTUs for same name are in OTU set then both have to be provisional
     # * missaplication (?)
     nil
+  end
+
+
+  # Name phrase is for appended phrases like senso stricto and senso lato
+  def self.name_phrase(otu, vocab_id)
+    da = DataAttribute.find_by(type: 'InternalAttribute',
+                               controlled_vocabulary_term_id: vocab_id,
+                               attribute_subject_id: otu.id)
+    da&.value
   end
 
   # The scrutinizer concept is unused at present
@@ -109,12 +118,16 @@ module Export::Coldp::Files::Taxon
     #   `taxon_name_id: nil`  - uniquify via Ruby hash keys
     observed_taxon_name_ids = { }
 
+    # TODO: optional Taxon.alternativeID field allows inclusion of external identifiers: https://github.com/CatalogueOfLife/coldp#alternativeid-1 https://github.com/CatalogueOfLife/coldp#identifiers
+    #   e.g., gbif:2704179,col:6W3C4,BOLD:AAJ2287,wikidata:Q157571
+
     CSV.generate(col_sep: "\t") do |csv|
 
       csv << %w{
         ID
         parentID
         nameID
+        namePhrase
         provisional
         accordingToID
         scrutinizer
@@ -124,14 +137,13 @@ module Export::Coldp::Files::Taxon
         extinct
         temporalRangeStart
         temporalRangeEnd
-        lifezone
+        environment
         link
         remarks
       }
 
       taxon_remarks_vocab_id = Predicate.find_by(uri: 'https://github.com/catalogueoflife/coldp#Taxon.remarks',
                                                  project_id: otus[0]&.project_id)&.id
-
       otus.each do |o|
         # !! When a name is a synonmy (combination), but that combination has no OTU
         # !! then the parent of the name in the taxon table is nil
@@ -167,19 +179,20 @@ module Export::Coldp::Files::Taxon
         parent_id = (root_otu_id == o.id ? nil : parent_id )
 
         csv << [
-          o.id,                                      # ID (Taxon)
-          parent_id,                                 # parentID (Taxon)
-          o.taxon_name.id,                           # nameID (Name)
-          provisional(o),                            # provisional
-          according_to_id(o),                        # accordingToID
-          scrutinizer(o),                            # scrutinizer
-          scrutinizer_id(o),                         # scrutinizerID
-          scrutinizer_date(o),                       # scrutizinerDate
-          reference_id(sources),                     # referenceID
+          o.id,                                              # ID (Taxon)
+          parent_id,                                         # parentID (Taxon)
+          o.taxon_name.id,                                   # nameID (Name)
+          name_phrase(o, name_phrase_vocab_id),              # namePhrase
+          provisional(o),                                    # provisional
+          according_to_id(o),                                # accordingToID
+          scrutinizer(o),                                    # scrutinizer
+          scrutinizer_id(o),                                 # scrutinizerID
+          scrutinizer_date(o),                               # scrutizinerDate
+          reference_id(sources),                             # referenceID
           predicate_value(o, :extinct),              # extinct
           predicate_value(o, :temporal_range_start), # temporalRangeStart
           predicate_value(o, :temporal_range_end),   # temporalRangeEnd
-          predicate_value(o, :lifezone),             # lifezone
+          predicate_value(o, :lifezone),             # environment (formerly named lifezone)
           link(o),                                   # link
           remarks(o, taxon_remarks_vocab_id)         # remarks
         ]