+++ lib/queries/query.rb
@@ -190,13 +190,23 @@ module Queries
     #   table is defined in each query, it is the class of instances being returned
 
     # params attribute [Symbol]
-    #   a facet for use when params include, `author`, and `exact_author` combinations
+    #   a facet for use when params include `author`, and `exact_author` pattern combinations
     #   See queries/source/filter.rb for example use
+    #   See /spec/lib/queries/source/filter_spec.rb
+    #  !! Whitespace (e.g. tabs, newlines) is ignored when exact is used !!!
     def attribute_exact_facet(attribute = nil)
       a = attribute.to_sym
       return nil if send(a).blank?
       if send("exact_#{a}".to_sym)
-        table[a].eq(send(a).strip)
+
+        v = send(a)
+        v.gsub!(/\s+/, ' ')
+        v = ::Regexp.escape(v)
+        v.gsub!(/\\\s+/, '\s*') # spaces are escaped, but we need to expand them in case we dont' get them caught
+        v = '^\s*' + v + '\s*$'
+
+        # !! May need to add table name here preceeding a.
+        Arel::Nodes::SqlLiteral.new( "#{a} ~ '#{v}'" )
       else
         table[a].matches('%' + send(a).strip.gsub(/\s+/, '%') + '%')
       end
@@ -384,12 +394,12 @@ module Queries
 
     def autocomplete_common_name_exact
       return nil if no_terms?
-      query_base.joins(:common_names).where(common_name_name.to_sql).limit(1)
+      base_query.joins(:common_names).where(common_name_name.to_sql).limit(1)
     end
 
     def autocomplete_common_name_like
       return nil if no_terms?
-      query_base.joins(:common_names).where(common_name_wild_pieces.to_sql).limit(5)
+      base_query.joins(:common_names).where(common_name_wild_pieces.to_sql).limit(5)
     end
 
   end